% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFuncs.R
\docType{data}
\name{.parms}
\alias{.parms}
\title{Set the package's parameters}
\format{An object of class \code{list} of length 5.}
\usage{
.parms
}
\description{
The function 'getChunkSize' and 'setChunkSize' can be used to get or set the chunk size
}
\examples{
.parms$getChunkSize()
.parms$setChunkSize()

}
\keyword{datasets}
