% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSparse.R
\name{openSparse}
\alias{openSparse}
\title{Create an 'openSparse' object}
\usage{
openSparse(file)
}
\arguments{
\item{file}{character(1) path to hdf5 file containing a CSC formated sparse matrix.}
}
\value{
An object of class 'openSparse'
}
\description{
The 'openSparse' object provide a variaty of matrix operation on the sparse data.
}
\examples{
library(rhdf5)
library(ExperimentHub)
library(HDF5Array)
library("tictoc")
hub=ExperimentHub::ExperimentHub()
fl=hub[["EH1039"]]
mydata=openSparse(fl)

}
