% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSparse.R
\name{openSparse}
\alias{openSparse}
\title{Create an 'openSparse' object}
\usage{
openSparse(file)
}
\arguments{
\item{file}{character(1) path to hdf5 file containing ....}
}
\value{
An object of class 'openSparse'
}
\description{
This is description
}
\examples{
fl <- "..."
openSparse(fl)

}
